chore: switch from `phantomjs2-ext` to regular `phantomjs2` dependency
The standard `phantomjs2` dependency now works for Windows and has support
for the "official" final PhantomJS 2.0.0 release (`phantomjs2-ext` relies
on custom builds and still points to a pre release of 2.0). Due to an `npm`
issue we're currently pointing to someone's PR that fixes a problem with
`phantomjs2` being installed via `npm` but will point to a proper semver
when the PR is accepted and appears in a regular release.